+++ Coursework/Coursework/LightShader.cpp
@@ -85,7 +85,7 @@ void LightShader::initShader(const wchar_t* vsFilename, const wchar_t* psFilenam
 
 
 void LightShader::setShaderParameters(ID3D11DeviceContext* deviceContext, const XMMATRIX& worldMatrix, const XMMATRIX& viewMatrix, const XMMATRIX& projectionMatrix,
-									  int lightCount, const SceneLight* lights, Camera* camera, const Material* mat)
+									size_t lightCount, const SceneLight* lights, Camera* camera, const Material* mat)
 {
 	HRESULT result;
 	D3D11_MAPPED_SUBRESOURCE mappedResource;
